データ型変更

+++ db/migrate/20231207061711_create_relationships.rb
@@ -1,8 +1,8 @@
 class CreateRelationships < ActiveRecord::Migration[6.1]
   def change
     create_table :relationships do |t|
-      t.integer :followed_id, null: false, foreign_key: { to_table: :users }
-      t.integer :follower_id, null: false, foreign_key: { to_table: :users }
+      t.bigint :followed_id, null: false, foreign_key: { to_table: :users }
+      t.bigint :follower_id, null: false, foreign_key: { to_table: :users }
 
       t.timestamps
     end
